BC-167 Adjust translations, move components, fix typos
