+++ locale/de.json
@@ -200,7 +200,6 @@
 	"global.sidebar.tasks.tasksCurrent": "Aktuelle Aufgaben",
 	"global.sidebar.teacher": "Lehrkräfte",
 	"global.sidebar.teams": "Teams",
-	"global.topbar.actions.contactAdmin": "Admin deiner Schule kontaktieren",
 	"global.topbar.actions.contactSupport": "Wunsch oder Problem senden",
 	"global.topbar.actions.help": "Hilfebereich",
 	"global.topbar.actions.training": "Fortbildungen",
