+++ locale/en.json
@@ -199,10 +199,9 @@
 	"global.sidebar.tasksUngraded": "Unchecked",
 	"global.sidebar.teacher": "Teacher",
 	"global.sidebar.teams": "Teams",
-	"global.topbar.actions.contactAdmin": "Contact school admin",
-	"global.topbar.actions.contactSupport": "Report errors or requests",
+	"global.topbar.actions.contactSupport": "Send request or problem",
 	"global.topbar.actions.help": "Help section",
-	"global.topbar.actions.training": "Training",
+	"global.topbar.actions.training": "Advanced trainings",
 	"global.topbar.settings": "Settings",
 	"global.topbar.roleName.administrator": "Administrator",
 	"global.topbar.roleName.demo": "Demo",
