+++ locale/es.json
@@ -199,10 +199,9 @@
 	"global.sidebar.tasksUngraded": "Sin auditar",
 	"global.sidebar.teacher": "Profesor",
 	"global.sidebar.teams": "Equipos",
-	"global.topbar.actions.contactAdmin": "Contacta a al admin de tu escuela",
-	"global.topbar.actions.contactSupport": "Reporta errores o sugerencias",
+	"global.topbar.actions.contactSupport": "Enviar solicitud o problema",
 	"global.topbar.actions.help": "Sección de ayuda",
-	"global.topbar.actions.training": "Capacitación",
+	"global.topbar.actions.training": "Formaciones avanzadas",
 	"global.topbar.settings": "Configuración",
 	"global.topbar.roleName.administrator": "Administrador",
 	"global.topbar.roleName.demo": "Demostración",
