+++ src/pages/administration/students/index.vue
@@ -144,7 +144,7 @@
 import SchoolsModule from "@/store/schools";
 import { mapGetters } from "vuex";
 import EnvConfigModule from "@/store/env-config";
-import DefaultWireframe from "@components/molecules/DefaultWireframe.vue";
+import DefaultWireframe from "@components/templates/DefaultWireframe.vue";
 import BackendDataTable from "@components/organisms/DataTable/BackendDataTable";
 import FabFloating from "@components/molecules/FabFloating";
 import DataFilter from "@components/organisms/DataFilter/DataFilter";
